[nextjs-api] api/maps/tky2jgdg, jgd2tkyg を express.js と共通化し next.js に実装
